software: fix drawseg xrange item coord type being too small
both sprite and drawseg coords are int32
might fix some clipping issues

+++ src/r_things.cpp
@@ -70,7 +70,7 @@ INT16 *screenheightarray = NULL;
 
 typedef struct drawseg_xrange_item_s
 {
-	INT16 x1, x2;
+	INT32 x1, x2;
 	drawseg_t *user;
 } drawseg_xrange_item_t;
